+++ src/main/java/com/frejdh/util/common/ansi/models/AnsiCodeInterface.java
@@ -1,4 +1,4 @@
-package com.frejdh.util.common.ansi;
+package com.frejdh.util.common.ansi.models;
 
 /**
  * Interface for returning the code for the color/style etc.